chat: don't draw half line (lines shifted when active)

+++ src/main/java/de/bixilon/minosoft/gui/rendering/gui/elements/text/TextFlowElement.kt
@@ -102,7 +102,7 @@ open class TextFlowElement(
     override fun forceSilentApply() {
         val visibleLines: MutableList<TextFlowLineElement> = mutableListOf()
         val maxSize = maxSize
-        val maxLines = maxSize.y / (properties.lineHeight + properties.lineSpacing)
+        val maxLines = (maxSize.y / (properties.lineHeight + properties.lineSpacing)).toInt()
         val currentTime = millis()
         var textSize = Vec2.EMPTY
         val active = this.active
